@@ -19,7 +19,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: '>=3.4.1 <4.0.0'
+  sdk: ^3.5.0
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -40,30 +40,30 @@ dependencies:
   flash: ^3.1.1
   flutter:
     sdk: flutter
-  flutter_form_builder: ^9.3.0
+  flutter_form_builder: ^9.4.1
   flutter_hooks: ^0.20.5
   flutter_secure_storage: ^9.2.2
   flutter_staggered_grid_view: ^0.7.0
   form_builder_validators: ^11.0.0
   freezed_annotation: ^2.4.4
-  go_router: ^14.2.1
+  go_router: ^14.2.3
   hooks_riverpod: ^2.5.1
   json_annotation: ^4.9.0
   path: ^1.9.0
-  path_provider: ^2.1.3
+  path_provider: ^2.1.4
   riverpod_annotation: ^2.3.5
   rxdart: ^0.28.0
-  sembast: ^3.7.1
-  shadcn_ui: ^0.6.3
-  shared_preferences: ^2.2.3
-  skeletonizer: ^1.4.1
+  sembast: ^3.7.2
+  shadcn_ui: ^0.8.0
+  shared_preferences: ^2.3.2
+  skeletonizer: ^1.4.2
   sliver_tools: ^0.2.12
   super_sliver_list: ^0.4.1
 
 dev_dependencies:
-  build_runner: ^2.4.11
+  build_runner: ^2.4.12
   chopper_generator: ^8.0.1
-  custom_lint: ^0.6.4
+  custom_lint: ^0.6.5
   enven: ^1.0.1
   flutter_test:
     sdk: flutter
