@@ -6,21 +6,22 @@ part of 'image_notifier.dart';
 // RiverpodGenerator
 // **************************************************************************
 
-String _$imagesNotifierHash() => r'330ff4388088ed2f267dbe54f5324ae3269ccf7d';
+String _$productImagesNotifierHash() =>
+    r'861497a68f7931261154c38c298e2b1ff2931afe';
 
-/// See also [ImagesNotifier].
-@ProviderFor(ImagesNotifier)
-final imagesNotifierProvider =
-    NotifierProvider<ImagesNotifier, ImageState>.internal(
-  ImagesNotifier.new,
-  name: r'imagesNotifierProvider',
+/// See also [ProductImagesNotifier].
+@ProviderFor(ProductImagesNotifier)
+final productImagesNotifierProvider =
+    NotifierProvider<ProductImagesNotifier, ProductImageState>.internal(
+  ProductImagesNotifier.new,
+  name: r'productImagesNotifierProvider',
   debugGetCreateSourceHash: const bool.fromEnvironment('dart.vm.product')
       ? null
-      : _$imagesNotifierHash,
+      : _$productImagesNotifierHash,
   dependencies: null,
   allTransitiveDependencies: null,
 );
 
-typedef _$ImagesNotifier = Notifier<ImageState>;
+typedef _$ProductImagesNotifier = Notifier<ProductImageState>;
 // ignore_for_file: type=lint
 // ignore_for_file: subtype_of_sealed_class, invalid_use_of_internal_member, invalid_use_of_visible_for_testing_member
