@@ -2,19 +2,21 @@ import 'dart:io';
 
 import 'package:chopper/chopper.dart';
 import 'package:classic_shop_admin/src/core/data/network_exceptions.dart';
+import 'package:classic_shop_admin/src/core/data/pagination_config.dart';
 import 'package:classic_shop_admin/src/core/data/remote_response.dart';
 import 'package:classic_shop_admin/src/core/data/response_headers.dart';
 import 'package:classic_shop_admin/src/core/data/response_headers_cache.dart';
 import 'package:classic_shop_admin/src/features/image/data/image_api.dart';
 import 'package:classic_shop_admin/src/features/image/data/image_dto.dart';
+import 'package:classic_shop_admin/src/features/image/helpers/enums.dart';
 import 'package:flutter/foundation.dart';
 
 abstract class IImageRemoteService {
-  Future<RemoteResponse<List<ImageKitDTO>>> fetchImages({
+  Future<RemoteResponse<List<ImageKitDTO>>> fetchImageKits({
     required int adminId,
     required Uri requestUri,
   });
-  Future<RemoteResponse<ImageDTO>> createImages({
+  Future<RemoteResponse<ProductImageDTO>> createImages({
     required int adminId,
     required String productImage1,
     required String productImage2,
@@ -25,21 +27,23 @@ abstract class IImageRemoteService {
 class ImageRemoteService implements IImageRemoteService {
   const ImageRemoteService(
     this._imageApi,
+    this._imageAdminApi,
     this._headersCache,
   );
 
   final ImageApi _imageApi;
+  final ImageAdminApi _imageAdminApi;
   final ResponseHeadersCache _headersCache;
 
   @override
-  Future<RemoteResponse<ImageDTO>> createImages({
+  Future<RemoteResponse<ProductImageDTO>> createImages({
     required int adminId,
     required String productImage1,
     required String productImage2,
     required String productImage3,
   }) async {
     try {
-      final response = await _imageApi.createProductImages(
+      final response = await _imageAdminApi.createProductImages(
         adminId: adminId.toString(),
         data: {
           'product_image_1': productImage1,
@@ -58,7 +62,7 @@ class ImageRemoteService implements IImageRemoteService {
         throw const RestApiException();
       }
 
-      final imageDTO = ImageDTO.fromJson(body);
+      final imageDTO = ProductImageDTO.fromJson(body);
 
       return RemoteResponse.withNewData(imageDTO, nextAvailable: false);
     } on SocketException {
@@ -67,7 +71,7 @@ class ImageRemoteService implements IImageRemoteService {
   }
 
   @override
-  Future<RemoteResponse<List<ImageKitDTO>>> fetchImages({
+  Future<RemoteResponse<List<ImageKitDTO>>> fetchImageKits({
     required int adminId,
     required Uri requestUri,
   }) async {
@@ -75,7 +79,7 @@ class ImageRemoteService implements IImageRemoteService {
     try {
       late final Response<List<Map<String, dynamic>>> response;
 
-      response = await _imageApi.listProductImagesKit(
+      response = await _imageAdminApi.listProductImagesKit(
         ifNoneMatch: previousHeaders?.etag ?? '',
         adminId: adminId.toString(),
       );
@@ -98,6 +102,62 @@ class ImageRemoteService implements IImageRemoteService {
         return RemoteResponse.withNewData(
           convertedBody.map(ImageKitDTO.fromJson).toList(),
           nextAvailable: false,
+        );
+      } else {
+        throw RestApiException(response.statusCode);
+      }
+    } on SocketException {
+      return const RemoteResponse.noConnection();
+    }
+  }
+
+  Future<RemoteResponse<List<ProductImageDTO>>> fetchImages({
+    required ImagesFunction imagesFunction,
+    required Uri requestUri,
+    int? lastImageId,
+    String? query,
+    int? pageSize,
+  }) async {
+    final previousHeaders = await _headersCache.getHeaders(requestUri);
+    try {
+      late final Response<List<Map<String, dynamic>>> response;
+      switch (imagesFunction) {
+        case ImagesFunction.getImages:
+          response = await _imageApi.getImagesV2(
+            ifNoneMatch: previousHeaders?.etag ?? '',
+            pageSize: pageSize ?? PaginationConfig.itemsPerPage,
+          );
+
+        case ImagesFunction.getImagesNextPage:
+          response = await _imageApi.getImagesNextPage(
+            ifNoneMatch: previousHeaders?.etag ?? '',
+            pageSize: pageSize ?? PaginationConfig.itemsPerPage,
+            lastImageId: lastImageId ?? 0,
+          );
+      }
+
+      if (response.statusCode == 304) {
+        return RemoteResponse.notModified(
+          nextAvailable: previousHeaders?.nextAvailable ?? false,
+        );
+      }
+
+      if (response.statusCode == 204) {
+        return const RemoteResponse.noContent();
+      }
+
+      if (response.statusCode == 200) {
+        if (response.body!.isEmpty) {
+          return const RemoteResponse.noContent();
+        }
+        final headers = ResponseHeaders.parse(response);
+        await _headersCache.saveHeaders(requestUri, headers);
+        // response as Response<List<Map<String, dynamic>>>;
+        final convertedBody =
+            response.body!.map(ProductImageDTO.fromJson).toList();
+        return RemoteResponse.withNewData(
+          convertedBody,
+          nextAvailable: headers.nextAvailable ?? true,
         );
       } else {
         throw RestApiException(response.statusCode);
