@@ -15,34 +15,35 @@ final _privateConstructorUsedError = UnsupportedError(
     'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more information: https://github.com/rrousselGit/freezed#adding-getters-and-methods-to-our-models');
 
 /// @nodoc
-mixin _$Image {
+mixin _$ImageKit {
   String get url => throw _privateConstructorUsedError;
 
-  /// Create a copy of Image
+  /// Create a copy of ImageKit
   /// with the given fields replaced by the non-null parameter values.
   @JsonKey(includeFromJson: false, includeToJson: false)
-  $ImageCopyWith<Image> get copyWith => throw _privateConstructorUsedError;
+  $ImageKitCopyWith<ImageKit> get copyWith =>
+      throw _privateConstructorUsedError;
 }
 
 /// @nodoc
-abstract class $ImageCopyWith<$Res> {
-  factory $ImageCopyWith(Image value, $Res Function(Image) then) =
-      _$ImageCopyWithImpl<$Res, Image>;
+abstract class $ImageKitCopyWith<$Res> {
+  factory $ImageKitCopyWith(ImageKit value, $Res Function(ImageKit) then) =
+      _$ImageKitCopyWithImpl<$Res, ImageKit>;
   @useResult
   $Res call({String url});
 }
 
 /// @nodoc
-class _$ImageCopyWithImpl<$Res, $Val extends Image>
-    implements $ImageCopyWith<$Res> {
-  _$ImageCopyWithImpl(this._value, this._then);
+class _$ImageKitCopyWithImpl<$Res, $Val extends ImageKit>
+    implements $ImageKitCopyWith<$Res> {
+  _$ImageKitCopyWithImpl(this._value, this._then);
 
   // ignore: unused_field
   final $Val _value;
   // ignore: unused_field
   final $Res Function($Val) _then;
 
-  /// Create a copy of Image
+  /// Create a copy of ImageKit
   /// with the given fields replaced by the non-null parameter values.
   @pragma('vm:prefer-inline')
   @override
@@ -59,31 +60,32 @@ class _$ImageCopyWithImpl<$Res, $Val extends Image>
 }
 
 /// @nodoc
-abstract class _$$ImageImplCopyWith<$Res> implements $ImageCopyWith<$Res> {
-  factory _$$ImageImplCopyWith(
-          _$ImageImpl value, $Res Function(_$ImageImpl) then) =
-      __$$ImageImplCopyWithImpl<$Res>;
+abstract class _$$ImageKitImplCopyWith<$Res>
+    implements $ImageKitCopyWith<$Res> {
+  factory _$$ImageKitImplCopyWith(
+          _$ImageKitImpl value, $Res Function(_$ImageKitImpl) then) =
+      __$$ImageKitImplCopyWithImpl<$Res>;
   @override
   @useResult
   $Res call({String url});
 }
 
 /// @nodoc
-class __$$ImageImplCopyWithImpl<$Res>
-    extends _$ImageCopyWithImpl<$Res, _$ImageImpl>
-    implements _$$ImageImplCopyWith<$Res> {
-  __$$ImageImplCopyWithImpl(
-      _$ImageImpl _value, $Res Function(_$ImageImpl) _then)
+class __$$ImageKitImplCopyWithImpl<$Res>
+    extends _$ImageKitCopyWithImpl<$Res, _$ImageKitImpl>
+    implements _$$ImageKitImplCopyWith<$Res> {
+  __$$ImageKitImplCopyWithImpl(
+      _$ImageKitImpl _value, $Res Function(_$ImageKitImpl) _then)
       : super(_value, _then);
 
-  /// Create a copy of Image
+  /// Create a copy of ImageKit
   /// with the given fields replaced by the non-null parameter values.
   @pragma('vm:prefer-inline')
   @override
   $Res call({
     Object? url = null,
   }) {
-    return _then(_$ImageImpl(
+    return _then(_$ImageKitImpl(
       url: null == url
           ? _value.url
           : url // ignore: cast_nullable_to_non_nullable
@@ -94,48 +96,395 @@ class __$$ImageImplCopyWithImpl<$Res>
 
 /// @nodoc
 
-class _$ImageImpl extends _Image {
-  const _$ImageImpl({required this.url}) : super._();
+class _$ImageKitImpl extends _ImageKit {
+  const _$ImageKitImpl({required this.url}) : super._();
 
   @override
   final String url;
 
   @override
   String toString() {
-    return 'Image(url: $url)';
+    return 'ImageKit(url: $url)';
   }
 
   @override
   bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
-            other is _$ImageImpl &&
+            other is _$ImageKitImpl &&
             (identical(other.url, url) || other.url == url));
   }
 
   @override
   int get hashCode => Object.hash(runtimeType, url);
 
-  /// Create a copy of Image
+  /// Create a copy of ImageKit
   /// with the given fields replaced by the non-null parameter values.
   @JsonKey(includeFromJson: false, includeToJson: false)
   @override
   @pragma('vm:prefer-inline')
-  _$$ImageImplCopyWith<_$ImageImpl> get copyWith =>
-      __$$ImageImplCopyWithImpl<_$ImageImpl>(this, _$identity);
+  _$$ImageKitImplCopyWith<_$ImageKitImpl> get copyWith =>
+      __$$ImageKitImplCopyWithImpl<_$ImageKitImpl>(this, _$identity);
 }
 
-abstract class _Image extends Image {
-  const factory _Image({required final String url}) = _$ImageImpl;
-  const _Image._() : super._();
+abstract class _ImageKit extends ImageKit {
+  const factory _ImageKit({required final String url}) = _$ImageKitImpl;
+  const _ImageKit._() : super._();
 
   @override
   String get url;
 
-  /// Create a copy of Image
+  /// Create a copy of ImageKit
   /// with the given fields replaced by the non-null parameter values.
   @override
   @JsonKey(includeFromJson: false, includeToJson: false)
-  _$$ImageImplCopyWith<_$ImageImpl> get copyWith =>
+  _$$ImageKitImplCopyWith<_$ImageKitImpl> get copyWith =>
+      throw _privateConstructorUsedError;
+}
+
+/// @nodoc
+mixin _$ProductImage {
+  int get id => throw _privateConstructorUsedError;
+  String get productImage_1 => throw _privateConstructorUsedError;
+  String get productImage_2 => throw _privateConstructorUsedError;
+  String get productImage_3 => throw _privateConstructorUsedError;
+
+  /// Create a copy of ProductImage
+  /// with the given fields replaced by the non-null parameter values.
+  @JsonKey(includeFromJson: false, includeToJson: false)
+  $ProductImageCopyWith<ProductImage> get copyWith =>
+      throw _privateConstructorUsedError;
+}
+
+/// @nodoc
+abstract class $ProductImageCopyWith<$Res> {
+  factory $ProductImageCopyWith(
+          ProductImage value, $Res Function(ProductImage) then) =
+      _$ProductImageCopyWithImpl<$Res, ProductImage>;
+  @useResult
+  $Res call(
+      {int id,
+      String productImage_1,
+      String productImage_2,
+      String productImage_3});
+}
+
+/// @nodoc
+class _$ProductImageCopyWithImpl<$Res, $Val extends ProductImage>
+    implements $ProductImageCopyWith<$Res> {
+  _$ProductImageCopyWithImpl(this._value, this._then);
+
+  // ignore: unused_field
+  final $Val _value;
+  // ignore: unused_field
+  final $Res Function($Val) _then;
+
+  /// Create a copy of ProductImage
+  /// with the given fields replaced by the non-null parameter values.
+  @pragma('vm:prefer-inline')
+  @override
+  $Res call({
+    Object? id = null,
+    Object? productImage_1 = null,
+    Object? productImage_2 = null,
+    Object? productImage_3 = null,
+  }) {
+    return _then(_value.copyWith(
+      id: null == id
+          ? _value.id
+          : id // ignore: cast_nullable_to_non_nullable
+              as int,
+      productImage_1: null == productImage_1
+          ? _value.productImage_1
+          : productImage_1 // ignore: cast_nullable_to_non_nullable
+              as String,
+      productImage_2: null == productImage_2
+          ? _value.productImage_2
+          : productImage_2 // ignore: cast_nullable_to_non_nullable
+              as String,
+      productImage_3: null == productImage_3
+          ? _value.productImage_3
+          : productImage_3 // ignore: cast_nullable_to_non_nullable
+              as String,
+    ) as $Val);
+  }
+}
+
+/// @nodoc
+abstract class _$$ProductImageImplCopyWith<$Res>
+    implements $ProductImageCopyWith<$Res> {
+  factory _$$ProductImageImplCopyWith(
+          _$ProductImageImpl value, $Res Function(_$ProductImageImpl) then) =
+      __$$ProductImageImplCopyWithImpl<$Res>;
+  @override
+  @useResult
+  $Res call(
+      {int id,
+      String productImage_1,
+      String productImage_2,
+      String productImage_3});
+}
+
+/// @nodoc
+class __$$ProductImageImplCopyWithImpl<$Res>
+    extends _$ProductImageCopyWithImpl<$Res, _$ProductImageImpl>
+    implements _$$ProductImageImplCopyWith<$Res> {
+  __$$ProductImageImplCopyWithImpl(
+      _$ProductImageImpl _value, $Res Function(_$ProductImageImpl) _then)
+      : super(_value, _then);
+
+  /// Create a copy of ProductImage
+  /// with the given fields replaced by the non-null parameter values.
+  @pragma('vm:prefer-inline')
+  @override
+  $Res call({
+    Object? id = null,
+    Object? productImage_1 = null,
+    Object? productImage_2 = null,
+    Object? productImage_3 = null,
+  }) {
+    return _then(_$ProductImageImpl(
+      id: null == id
+          ? _value.id
+          : id // ignore: cast_nullable_to_non_nullable
+              as int,
+      productImage_1: null == productImage_1
+          ? _value.productImage_1
+          : productImage_1 // ignore: cast_nullable_to_non_nullable
+              as String,
+      productImage_2: null == productImage_2
+          ? _value.productImage_2
+          : productImage_2 // ignore: cast_nullable_to_non_nullable
+              as String,
+      productImage_3: null == productImage_3
+          ? _value.productImage_3
+          : productImage_3 // ignore: cast_nullable_to_non_nullable
+              as String,
+    ));
+  }
+}
+
+/// @nodoc
+
+class _$ProductImageImpl extends _ProductImage {
+  const _$ProductImageImpl(
+      {required this.id,
+      required this.productImage_1,
+      required this.productImage_2,
+      required this.productImage_3})
+      : super._();
+
+  @override
+  final int id;
+  @override
+  final String productImage_1;
+  @override
+  final String productImage_2;
+  @override
+  final String productImage_3;
+
+  @override
+  String toString() {
+    return 'ProductImage(id: $id, productImage_1: $productImage_1, productImage_2: $productImage_2, productImage_3: $productImage_3)';
+  }
+
+  @override
+  bool operator ==(Object other) {
+    return identical(this, other) ||
+        (other.runtimeType == runtimeType &&
+            other is _$ProductImageImpl &&
+            (identical(other.id, id) || other.id == id) &&
+            (identical(other.productImage_1, productImage_1) ||
+                other.productImage_1 == productImage_1) &&
+            (identical(other.productImage_2, productImage_2) ||
+                other.productImage_2 == productImage_2) &&
+            (identical(other.productImage_3, productImage_3) ||
+                other.productImage_3 == productImage_3));
+  }
+
+  @override
+  int get hashCode => Object.hash(
+      runtimeType, id, productImage_1, productImage_2, productImage_3);
+
+  /// Create a copy of ProductImage
+  /// with the given fields replaced by the non-null parameter values.
+  @JsonKey(includeFromJson: false, includeToJson: false)
+  @override
+  @pragma('vm:prefer-inline')
+  _$$ProductImageImplCopyWith<_$ProductImageImpl> get copyWith =>
+      __$$ProductImageImplCopyWithImpl<_$ProductImageImpl>(this, _$identity);
+}
+
+abstract class _ProductImage extends ProductImage {
+  const factory _ProductImage(
+      {required final int id,
+      required final String productImage_1,
+      required final String productImage_2,
+      required final String productImage_3}) = _$ProductImageImpl;
+  const _ProductImage._() : super._();
+
+  @override
+  int get id;
+  @override
+  String get productImage_1;
+  @override
+  String get productImage_2;
+  @override
+  String get productImage_3;
+
+  /// Create a copy of ProductImage
+  /// with the given fields replaced by the non-null parameter values.
+  @override
+  @JsonKey(includeFromJson: false, includeToJson: false)
+  _$$ProductImageImplCopyWith<_$ProductImageImpl> get copyWith =>
+      throw _privateConstructorUsedError;
+}
+
+/// @nodoc
+mixin _$ProductImageUrl {
+  int get id => throw _privateConstructorUsedError;
+  String get url => throw _privateConstructorUsedError;
+
+  /// Create a copy of ProductImageUrl
+  /// with the given fields replaced by the non-null parameter values.
+  @JsonKey(includeFromJson: false, includeToJson: false)
+  $ProductImageUrlCopyWith<ProductImageUrl> get copyWith =>
+      throw _privateConstructorUsedError;
+}
+
+/// @nodoc
+abstract class $ProductImageUrlCopyWith<$Res> {
+  factory $ProductImageUrlCopyWith(
+          ProductImageUrl value, $Res Function(ProductImageUrl) then) =
+      _$ProductImageUrlCopyWithImpl<$Res, ProductImageUrl>;
+  @useResult
+  $Res call({int id, String url});
+}
+
+/// @nodoc
+class _$ProductImageUrlCopyWithImpl<$Res, $Val extends ProductImageUrl>
+    implements $ProductImageUrlCopyWith<$Res> {
+  _$ProductImageUrlCopyWithImpl(this._value, this._then);
+
+  // ignore: unused_field
+  final $Val _value;
+  // ignore: unused_field
+  final $Res Function($Val) _then;
+
+  /// Create a copy of ProductImageUrl
+  /// with the given fields replaced by the non-null parameter values.
+  @pragma('vm:prefer-inline')
+  @override
+  $Res call({
+    Object? id = null,
+    Object? url = null,
+  }) {
+    return _then(_value.copyWith(
+      id: null == id
+          ? _value.id
+          : id // ignore: cast_nullable_to_non_nullable
+              as int,
+      url: null == url
+          ? _value.url
+          : url // ignore: cast_nullable_to_non_nullable
+              as String,
+    ) as $Val);
+  }
+}
+
+/// @nodoc
+abstract class _$$ProductImageUrlImplCopyWith<$Res>
+    implements $ProductImageUrlCopyWith<$Res> {
+  factory _$$ProductImageUrlImplCopyWith(_$ProductImageUrlImpl value,
+          $Res Function(_$ProductImageUrlImpl) then) =
+      __$$ProductImageUrlImplCopyWithImpl<$Res>;
+  @override
+  @useResult
+  $Res call({int id, String url});
+}
+
+/// @nodoc
+class __$$ProductImageUrlImplCopyWithImpl<$Res>
+    extends _$ProductImageUrlCopyWithImpl<$Res, _$ProductImageUrlImpl>
+    implements _$$ProductImageUrlImplCopyWith<$Res> {
+  __$$ProductImageUrlImplCopyWithImpl(
+      _$ProductImageUrlImpl _value, $Res Function(_$ProductImageUrlImpl) _then)
+      : super(_value, _then);
+
+  /// Create a copy of ProductImageUrl
+  /// with the given fields replaced by the non-null parameter values.
+  @pragma('vm:prefer-inline')
+  @override
+  $Res call({
+    Object? id = null,
+    Object? url = null,
+  }) {
+    return _then(_$ProductImageUrlImpl(
+      id: null == id
+          ? _value.id
+          : id // ignore: cast_nullable_to_non_nullable
+              as int,
+      url: null == url
+          ? _value.url
+          : url // ignore: cast_nullable_to_non_nullable
+              as String,
+    ));
+  }
+}
+
+/// @nodoc
+
+class _$ProductImageUrlImpl extends _ProductImageUrl {
+  const _$ProductImageUrlImpl({required this.id, required this.url})
+      : super._();
+
+  @override
+  final int id;
+  @override
+  final String url;
+
+  @override
+  String toString() {
+    return 'ProductImageUrl(id: $id, url: $url)';
+  }
+
+  @override
+  bool operator ==(Object other) {
+    return identical(this, other) ||
+        (other.runtimeType == runtimeType &&
+            other is _$ProductImageUrlImpl &&
+            (identical(other.id, id) || other.id == id) &&
+            (identical(other.url, url) || other.url == url));
+  }
+
+  @override
+  int get hashCode => Object.hash(runtimeType, id, url);
+
+  /// Create a copy of ProductImageUrl
+  /// with the given fields replaced by the non-null parameter values.
+  @JsonKey(includeFromJson: false, includeToJson: false)
+  @override
+  @pragma('vm:prefer-inline')
+  _$$ProductImageUrlImplCopyWith<_$ProductImageUrlImpl> get copyWith =>
+      __$$ProductImageUrlImplCopyWithImpl<_$ProductImageUrlImpl>(
+          this, _$identity);
+}
+
+abstract class _ProductImageUrl extends ProductImageUrl {
+  const factory _ProductImageUrl(
+      {required final int id,
+      required final String url}) = _$ProductImageUrlImpl;
+  const _ProductImageUrl._() : super._();
+
+  @override
+  int get id;
+  @override
+  String get url;
+
+  /// Create a copy of ProductImageUrl
+  /// with the given fields replaced by the non-null parameter values.
+  @override
+  @JsonKey(includeFromJson: false, includeToJson: false)
+  _$$ProductImageUrlImplCopyWith<_$ProductImageUrlImpl> get copyWith =>
       throw _privateConstructorUsedError;
 }
