@@ -21,7 +21,7 @@ final imageLocalServiceProvider = Provider<ImageLocalService>.internal(
 );
 
 typedef ImageLocalServiceRef = ProviderRef<ImageLocalService>;
-String _$imageApiHash() => r'fdbca36f43a5e33706d3932413de6d4b13ee38a1';
+String _$imageApiHash() => r'87bb6dcc77c5a53e88f62e273afe270ea1f2786f';
 
 /// See also [imageApi].
 @ProviderFor(imageApi)
@@ -35,8 +35,23 @@ final imageApiProvider = Provider<ImageApi>.internal(
 );
 
 typedef ImageApiRef = ProviderRef<ImageApi>;
+String _$imageAdminApiHash() => r'de231ce83f5b995bfcd823cbe7f04a11c14f1645';
+
+/// See also [imageAdminApi].
+@ProviderFor(imageAdminApi)
+final imageAdminApiProvider = Provider<ImageAdminApi>.internal(
+  imageAdminApi,
+  name: r'imageAdminApiProvider',
+  debugGetCreateSourceHash: const bool.fromEnvironment('dart.vm.product')
+      ? null
+      : _$imageAdminApiHash,
+  dependencies: null,
+  allTransitiveDependencies: null,
+);
+
+typedef ImageAdminApiRef = ProviderRef<ImageAdminApi>;
 String _$imageRemoteServiceHash() =>
-    r'ae92f18239f2ddf88ef5d371efde7e666466e1b2';
+    r'd82ac68f020416fc04890ab6047fdc0ddc01b1f4';
 
 /// See also [imageRemoteService].
 @ProviderFor(imageRemoteService)
@@ -51,7 +66,7 @@ final imageRemoteServiceProvider = Provider<ImageRemoteService>.internal(
 );
 
 typedef ImageRemoteServiceRef = ProviderRef<ImageRemoteService>;
-String _$imageRepositoryHash() => r'57adcf826cff77124c263f6ceb57385d364c8fe6';
+String _$imageRepositoryHash() => r'cbd2ad1d0de707be266b995b9b43ef7862ea381c';
 
 /// See also [imageRepository].
 @ProviderFor(imageRepository)
