@@ -28,6 +28,7 @@ mixin _$ProductDTO {
   bool get active => throw _privateConstructorUsedError;
   DateTime get createdAt => throw _privateConstructorUsedError;
   DateTime get updatedAt => throw _privateConstructorUsedError;
+  bool get nextAvailable => throw _privateConstructorUsedError;
 
   /// Serializes this ProductDTO to a JSON map.
   Map<String, dynamic> toJson() => throw _privateConstructorUsedError;
@@ -53,7 +54,8 @@ abstract class $ProductDTOCopyWith<$Res> {
       int brandId,
       bool active,
       DateTime createdAt,
-      DateTime updatedAt});
+      DateTime updatedAt,
+      bool nextAvailable});
 }
 
 /// @nodoc
@@ -79,6 +81,7 @@ class _$ProductDTOCopyWithImpl<$Res, $Val extends ProductDTO>
     Object? active = null,
     Object? createdAt = null,
     Object? updatedAt = null,
+    Object? nextAvailable = null,
   }) {
     return _then(_value.copyWith(
       id: null == id
@@ -113,6 +116,10 @@ class _$ProductDTOCopyWithImpl<$Res, $Val extends ProductDTO>
           ? _value.updatedAt
           : updatedAt // ignore: cast_nullable_to_non_nullable
               as DateTime,
+      nextAvailable: null == nextAvailable
+          ? _value.nextAvailable
+          : nextAvailable // ignore: cast_nullable_to_non_nullable
+              as bool,
     ) as $Val);
   }
 }
@@ -133,7 +140,8 @@ abstract class _$$ProductDTOImplCopyWith<$Res>
       int brandId,
       bool active,
       DateTime createdAt,
-      DateTime updatedAt});
+      DateTime updatedAt,
+      bool nextAvailable});
 }
 
 /// @nodoc
@@ -157,6 +165,7 @@ class __$$ProductDTOImplCopyWithImpl<$Res>
     Object? active = null,
     Object? createdAt = null,
     Object? updatedAt = null,
+    Object? nextAvailable = null,
   }) {
     return _then(_$ProductDTOImpl(
       id: null == id
@@ -191,6 +200,10 @@ class __$$ProductDTOImplCopyWithImpl<$Res>
           ? _value.updatedAt
           : updatedAt // ignore: cast_nullable_to_non_nullable
               as DateTime,
+      nextAvailable: null == nextAvailable
+          ? _value.nextAvailable
+          : nextAvailable // ignore: cast_nullable_to_non_nullable
+              as bool,
     ));
   }
 }
@@ -206,7 +219,8 @@ class _$ProductDTOImpl extends _ProductDTO {
       required this.brandId,
       required this.active,
       required this.createdAt,
-      required this.updatedAt})
+      required this.updatedAt,
+      required this.nextAvailable})
       : super._();
 
   factory _$ProductDTOImpl.fromJson(Map<String, dynamic> json) =>
@@ -228,10 +242,12 @@ class _$ProductDTOImpl extends _ProductDTO {
   final DateTime createdAt;
   @override
   final DateTime updatedAt;
+  @override
+  final bool nextAvailable;
 
   @override
   String toString() {
-    return 'ProductDTO(id: $id, name: $name, description: $description, categoryId: $categoryId, brandId: $brandId, active: $active, createdAt: $createdAt, updatedAt: $updatedAt)';
+    return 'ProductDTO(id: $id, name: $name, description: $description, categoryId: $categoryId, brandId: $brandId, active: $active, createdAt: $createdAt, updatedAt: $updatedAt, nextAvailable: $nextAvailable)';
   }
 
   @override
@@ -250,13 +266,15 @@ class _$ProductDTOImpl extends _ProductDTO {
             (identical(other.createdAt, createdAt) ||
                 other.createdAt == createdAt) &&
             (identical(other.updatedAt, updatedAt) ||
-                other.updatedAt == updatedAt));
+                other.updatedAt == updatedAt) &&
+            (identical(other.nextAvailable, nextAvailable) ||
+                other.nextAvailable == nextAvailable));
   }
 
   @JsonKey(includeFromJson: false, includeToJson: false)
   @override
   int get hashCode => Object.hash(runtimeType, id, name, description,
-      categoryId, brandId, active, createdAt, updatedAt);
+      categoryId, brandId, active, createdAt, updatedAt, nextAvailable);
 
   /// Create a copy of ProductDTO
   /// with the given fields replaced by the non-null parameter values.
@@ -283,7 +301,8 @@ abstract class _ProductDTO extends ProductDTO {
       required final int brandId,
       required final bool active,
       required final DateTime createdAt,
-      required final DateTime updatedAt}) = _$ProductDTOImpl;
+      required final DateTime updatedAt,
+      required final bool nextAvailable}) = _$ProductDTOImpl;
   const _ProductDTO._() : super._();
 
   factory _ProductDTO.fromJson(Map<String, dynamic> json) =
@@ -305,6 +324,8 @@ abstract class _ProductDTO extends ProductDTO {
   DateTime get createdAt;
   @override
   DateTime get updatedAt;
+  @override
+  bool get nextAvailable;
 
   /// Create a copy of ProductDTO
   /// with the given fields replaced by the non-null parameter values.
