@@ -6,6 +6,36 @@ part of 'provider.dart';
 // RiverpodGenerator
 // **************************************************************************
 
+String _$productLocalServiceHash() =>
+    r'f217ff94bbe5fe2a2b993ba8f4416e0f8241e7ee';
+
+/// See also [productLocalService].
+@ProviderFor(productLocalService)
+final productLocalServiceProvider = Provider<ProductLocalService>.internal(
+  productLocalService,
+  name: r'productLocalServiceProvider',
+  debugGetCreateSourceHash: const bool.fromEnvironment('dart.vm.product')
+      ? null
+      : _$productLocalServiceHash,
+  dependencies: null,
+  allTransitiveDependencies: null,
+);
+
+typedef ProductLocalServiceRef = ProviderRef<ProductLocalService>;
+String _$productApiHash() => r'5dbec4d9a1daece408b7e34954158528dfeb3acd';
+
+/// See also [productApi].
+@ProviderFor(productApi)
+final productApiProvider = Provider<ProductApi>.internal(
+  productApi,
+  name: r'productApiProvider',
+  debugGetCreateSourceHash:
+      const bool.fromEnvironment('dart.vm.product') ? null : _$productApiHash,
+  dependencies: null,
+  allTransitiveDependencies: null,
+);
+
+typedef ProductApiRef = ProviderRef<ProductApi>;
 String _$productAdminApiHash() => r'd8b5f971c4a029e779013fb4f33b880c6f327715';
 
 /// See also [productAdminApi].
@@ -22,7 +52,7 @@ final productAdminApiProvider = Provider<ProductAdminApi>.internal(
 
 typedef ProductAdminApiRef = ProviderRef<ProductAdminApi>;
 String _$productRemoteServiceHash() =>
-    r'7c0fac94e29b9cb058cf0eb61edcd446ff0b0911';
+    r'38bde31c4cad9a1bc757632b7e49ec41d6d91052';
 
 /// See also [productRemoteService].
 @ProviderFor(productRemoteService)
@@ -37,7 +67,7 @@ final productRemoteServiceProvider = Provider<ProductRemoteService>.internal(
 );
 
 typedef ProductRemoteServiceRef = ProviderRef<ProductRemoteService>;
-String _$productRepositoryHash() => r'7434d02c648172110c9d47d212731fd1c43ef00a';
+String _$productRepositoryHash() => r'57f94eca77cd85b00ef407bc91cfffe6d96838ea';
 
 /// See also [productRepository].
 @ProviderFor(productRepository)
