@@ -1,5 +1,5 @@
 import 'package:classic_shop_admin/src/features/product_items/core/shared/providers.dart';
-import 'package:classic_shop_admin/src/features/product_items/searched_products/presentation/widgets/searched_product_card.dart';
+import 'package:classic_shop_admin/src/features/product_items/searched_product_items/presentation/widgets/searched_product_card.dart';
 import 'package:classic_shop_admin/src/helpers/riverpod_observer.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_staggered_grid_view/flutter_staggered_grid_view.dart';
